Promjenjeni tipovi atributa spola i cilja mase. #49 - Promjenjeni cilj mase iz Integer u Float - Kreirana funkcija za kreiranje korisnika na web servisu - Izmjenjena vrijednost kod registracije za Cilj mase u Float - Izmjenjen atribut iz RetroKorisnika u Float (ciljMase) - Izmjenjen spol u klasi Korisnik u String

+++ FitApp/registracija/src/main/java/com/example/registracija/Registracija.java
@@ -41,7 +41,7 @@ public class Registracija extends AppCompatActivity {
 
     Float height;
     Float weight;
-    Integer weightGoal;
+    Float weightGoal;
     Float weightGainLossGoal;
     String gender;
     String datumRodenja;
@@ -199,7 +199,7 @@ public class Registracija extends AppCompatActivity {
                         @Override
                         public void onClick(DialogInterface dialog, int which) {
                             uiPickWeightGoal.setText(""+numberPicker.getValue()+" kg");
-                            weightGoal=Integer.valueOf(numberPicker.getValue());
+                            weightGoal=Float.valueOf(numberPicker.getValue());
                         }
                     });
